fix: basic rules to ignore debugger pauses
the v8 debugger may pause due to reasons other than breakpoints.
this commit adds a check for both the pause's reason, as well as the number of breakpoints related to the pause.
if the pause's reason is not `other` (the value it gets due to breakpoints) then we will ignore the pause.
if the pause's context lacks breakpoints, then we will ignore the pause as well.

adding unittests to assert this behavior.