chore: extract state-related things to a state module from the transmitter
the transmitter will now be responsible only for actually transmitting the state.
we could take it a step further by letting `state` decide what goes into the transmission, but that's enough for the moment.
this will allow us to send more data easily without overloading the `transmitter` with state-related tasks (specifically - loaded packages)